Physicsnemo  kNN Utility (#1079)
* Add cuml knn utility

* Add knn utilities with optimized backend selection.  Primarily targeting
data pipes and preprocessing, this kNN is not differentiable on
the distances returned.  However, you could use the indexes returned
and select on the `points` tensor, and *that* is differentiable.

The knn op has several backends:
- torch (brute force)
- scipy.spatial.KDTree (cpu only)
- cuml.Neighbords (GPU only)

The default backend is "auto" which will dispatch correctly.

* Update changelog

* Fix docstring and warning messages

* Resolve PR feedback.

Adds quiet promotion of bf16 data type to fp32 for cuml and scipy,
since that's not natively supported.

* Fix type annotations

* Darn linter got me

* Reduce cuml version requirement

* Disable cpu compile test

* I disabled the wrong CPU test ...